+++ Code/Systems/ExistingBuildingSystem.cs
@@ -17,15 +17,36 @@ namespace PlopTheGrowables
     public partial class ExistingBuildingSystem : GameSystemBase
     {
         private EntityQuery _emptyQuery;
+        private EntityQuery _allLockedBuildingsQuery;
+        private EntityQuery _allUnlockedBuildingsQuery;
+
+        /// <summary>
+        /// Gets the active instance.
+        /// </summary>
+        public static ExistingBuildingSystem Instance { get; private set; }
+
+        /// <summary>
+        /// Applies level-locking to all eligible buildings.
+        /// </summary>
+        internal void LockAllBuildings() => EntityManager.AddComponent<LevelLocked>(_allUnlockedBuildingsQuery);
+
+        /// <summary>
+        /// Removes level-locking from all eligible buildings.
+        /// </summary>
+        internal void UnlockAllBuildings() => EntityManager.RemoveComponent<LevelLocked>(_allLockedBuildingsQuery);
 
         /// <summary>
         /// Called when the system is created.
         /// </summary>
         protected override void OnCreate()
         {
+            Instance = this;
+
             base.OnCreate();
 
-            // Initialise query.
+            // Initialise queries.
+            _allLockedBuildingsQuery = SystemAPI.QueryBuilder().WithAll<Building, LevelLocked>().WithAny<ResidentialProperty, IndustrialProperty, CommercialProperty>().WithNone<Signature>().Build();
+            _allUnlockedBuildingsQuery = SystemAPI.QueryBuilder().WithAll<Building>().WithAny<ResidentialProperty, IndustrialProperty, CommercialProperty>().WithNone<Signature, LevelLocked>().Build();
             _emptyQuery = SystemAPI.QueryBuilder().WithAll<Building>().WithAny<ResidentialProperty, IndustrialProperty, CommercialProperty>().WithNone<Signature, PloppedBuilding, SpawnedBuilding>().Build();
             RequireForUpdate(_emptyQuery);
         }
@@ -38,6 +59,15 @@ namespace PlopTheGrowables
             // Set any existing uncategorised buildings as spawned.
             Mod.Instance.Log.Info($"Setting {_emptyQuery.CalculateEntityCount()} existing buildings as spawned.");
             EntityManager.AddComponent<SpawnedBuilding>(_emptyQuery);
+        }
+
+        /// <summary>
+        /// Called when the system is destroyed.
+        /// </summary>
+        protected override void OnDestroy()
+        {
+            Instance = null;
+            base.OnDestroy();
         }
     }
 }