+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add options panel buttons to lock or unlock all buildings.
+
 ### 0.4
 - Add option to automatically lock the levels of newly plopped buildings.
 - Add workaround for compatibility with Land Value Overhaul mod.